Convert ERA5 spectra to standard direction
ERA5 spectra by default use the 'going-to' convention for direction. This sets the default behaviour to be converting this to the wavespectra default of 'coming-from', but leaves the option to use 'going-to' if required.

I don't know if using 'going-to' will have any affect on other library functionality, if so perhaps 'coming-from' should simply be enforced, or a warning should be raised?